Break out TextLine into its own file.
